Scene template with working image brightness
There are two cameras in the scene. Scene Camera is just a renamed Main
Camera object. In addition to what it normally does, it captures the
image it sees into a render target.

This render target is applied to a quad placed somewhere else in the
scene.

A second camera points at this quad and the player will indirectly see
the game scene via this. Lighting changes are now applied to one image
instead of every object in the scene.

To change the lighting, go to Window -> Lighting -> Ambient Color and
pick a different value.

This scene works like any regular scenes, just make sure to hide Image
Camera somewhere far away and that the two cameras never overlap.