Enemy - Cone of Vision
Enemies will draw a ray to the player within the angle and distance
given.
If the ray hits the player, they have vision of the player.

The angle input should be half of what you want, since it is calculating
the angle that many degrees above and below from the forward vector.