Colliding Object Scene Load
When the player collides with an object, a scene will load. You can
specify which scene by modifying the code in the Player.cs file, under
the voidOnTriggerEnter2D function. Also, the object has a tag which
identifies it as an object that loads a scene, and some other object.

Additional Notes: Added a rigidbody2D to the player placeholder, to make
the collision work with the scene loading object. The scene loading
object is also instantiated within the scene view, for testing purposes,
but can be removed at any time.