Player Health Regen
Made a method to call a subroutine in order for health to regen over
time. Made a new variable called playerMaxHealth in order to stop
healing when we reach the max health. The amount of healing can be
changed amd its duration can be changed. The playerHealth variable was
made public for testing in the inspector. All variables can be changed
in the inspector.