Player Fuel Resource and Fuel Pickup
Allows player to pick up fuel for use. Has a maximum amount of fuel the
player can have in the Player.cs file and how the player gains fuel.
Most of the implementation is in the TractorBeamControls.cs, where the
fuel's value is decreased automatically unless the tractor beam is on
it. When the tractor beam is on it, it must stay for a certain amount of
seconds before it collects the fuel. In order to test this, you must
have a gameobject that has the "Fuel" tag.