Nebula Gel/ Ion Storm polish
Modified the player script to take into account the speed buffs from
nebula gels.

Modified the nebula gels to change a modifier in the player class
instead of directly effecting the velocity on entrance to the gels makes
for a smoother feel.

Also made it so that when debris and other objects enter the nebula gels
their speed will be buffed accordingly.

Lastly made changes to the damage formula for the pain nebulas so that
it does 25 damage every second gives time for the player to get out of
the pain nebula without instant death.