Pause button and fuel gauge
Changes:
- Added a test scene with just the HUD. You can use left/right arrows to
adjust the gauge's level here. The prefab in the folder is not linked
with the object in the scene and shouldn't be.
- Modified the Player script to add a public function to get the fuel
percentage.
- Modified the Level 1 scene to have the HUD in the HUD canvas.

Setup:
- Drop "Fuel Pause.prefab" into a UI canvas. The canvas needs to be set
up correctly. For the HUD canvas in Level 1, the following changes were
made:
Canvas -> Pixel Perfect: True
Canvas Scaler -> UI Scale Mode: Constant Pixel Size
- This breaks the hurt overlay (HurtEffectImage), so its Rect Transform
needs to be changed to the correct size.
- The player object needs to be linked to the Fuel Gauge script.
- (Optional) Clean up the Fuel Gauge script by moving the code in
TestPlayer() into Update() and remove everything related to
TestManual().