WallKnockBack semi success
Only able to bounce the player from from if touching down or left part
of the wall.
if touching up or right side, it will tend to try to bounce to the
downleft side.
Also did touch the Player script in order for the bounce effect to work.
Work in progress: try to make the player bounce if touching up or right
side of the wall