Implement Kill Counter For Player
Modified the EnemyCollision.cs script and Player.cs script. Made a
variable within the Player script to keep track of how many enemies it
has killed. Also made a method within it to increase the amount when it
gets called. Modified the EnemyCollision script so that when the enemy
dies, it will call the method within the Player script so that the kill
counter is increased.