SoundManager
Made a new scene named 'SoundManager'. Created a
'SongManagerPlaceholder'

Placeholder has two audio sources connected to it (a song by Taylor
Swift[Blank_Space] and a song by Drake[Energy]).

Referenced the two audios into a script named 'Song Manager'. It has a
function that if any key is pressed, two functions are called. One to
fade out the first audio and the other to fade in the latter.