Minimap
Created minimap with just a blue background and the character is
represented with a red arrow. Map is scaled to the size of the current
map. The Minimap is a camera with just the blue background. The arrow is
set to layer of player that I created so only the arrow is rendered
within the minimap camera.