Implementation for Player Failing Level
A specified scene is to be loaded when the player fails the level by
having its tethered object destroyed. Had to make an "Enemy" tag so that
the fail level implementation works. Also, created a prefab for the
Signal Booster object and created a new scene for testing the tether
implementation.