Task Finished
Move Textbox to the right then freeze player movement while text scrolls
through. while this is happening. the next level is then loaded. It gets
activated when the player presses a button.

I am having an issue with my display. For me, everything is centered,
but to others the textbox is off.